// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      events: {
        Row: {
          admin_id: string
          category: string
          created_at: string | null
          end_time: string
          event_name: string
          id: string
          number_of_choices: number
          number_of_items: number
          number_of_winners: number
          start_time: string
          status: Database["public"]["Enums"]["event_status"]
          updated_at: string | null
          vote_mode: Database["public"]["Enums"]["vote_mode"]
        }
        Insert: {
          admin_id: string
          category: string
          created_at?: string | null
          end_time: string
          event_name: string
          id?: string
          number_of_choices?: number
          number_of_items: number
          number_of_winners?: number
          start_time: string
          status?: Database["public"]["Enums"]["event_status"]
          updated_at?: string | null
          vote_mode?: Database["public"]["Enums"]["vote_mode"]
        }
        Update: {
          admin_id?: string
          category?: string
          created_at?: string | null
          end_time?: string
          event_name?: string
          id?: string
          number_of_choices?: number
          number_of_items?: number
          number_of_winners?: number
          start_time?: string
          status?: Database["public"]["Enums"]["event_status"]
          updated_at?: string | null
          vote_mode?: Database["public"]["Enums"]["vote_mode"]
        }
        Relationships: []
      }
      items: {
        Row: {
          created_at: string | null
          creator: string | null
          description: string | null
          event_id: string
          id: string
          image_url: string | null
          title: string
          votes: number | null
        }
        Insert: {
          created_at?: string | null
          creator?: string | null
          description?: string | null
          event_id: string
          id?: string
          image_url?: string | null
          title: string
          votes?: number | null
        }
        Update: {
          created_at?: string | null
          creator?: string | null
          description?: string | null
          event_id?: string
          id?: string
          image_url?: string | null
          title?: string
          votes?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "items_event_id_fkey"
            columns: ["event_id"]
            isOneToOne: false
            referencedRelation: "events"
            referencedColumns: ["id"]
          },
        ]
      }
      voters: {
        Row: {
          created_at: string | null
          event_id: string
          id: string
          used: boolean | null
          voted_at: string | null
          voter_code: string
        }
        Insert: {
          created_at?: string | null
          event_id: string
          id?: string
          used?: boolean | null
          voted_at?: string | null
          voter_code: string
        }
        Update: {
          created_at?: string | null
          event_id?: string
          id?: string
          used?: boolean | null
          voted_at?: string | null
          voter_code?: string
        }
        Relationships: [
          {
            foreignKeyName: "voters_event_id_fkey"
            columns: ["event_id"]
            isOneToOne: false
            referencedRelation: "events"
            referencedColumns: ["id"]
          },
        ]
      }
      votes: {
        Row: {
          blockchain_hash: string | null
          created_at: string | null
          event_id: string
          id: string
          item_ids: string[]
          voter_code: string
        }
        Insert: {
          blockchain_hash?: string | null
          created_at?: string | null
          event_id: string
          id?: string
          item_ids: string[]
          voter_code: string
        }
        Update: {
          blockchain_hash?: string | null
          created_at?: string | null
          event_id?: string
          id?: string
          item_ids?: string[]
          voter_code?: string
        }
        Relationships: [
          {
            foreignKeyName: "votes_event_id_fkey"
            columns: ["event_id"]
            isOneToOne: false
            referencedRelation: "events"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      event_status: "draft" | "active" | "paused" | "ended"
      vote_mode: "single" | "multiple"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never